+++ src/components/Sections/About/About.tsx
@@ -1,4 +1,4 @@
-import "./Styles.scss";
+import "./About.scss";
 import TagCloud from "TagCloud";
 import { useEffect } from "react";
 
@@ -58,33 +58,35 @@ const About: React.FC = () => {
             data-uk-grid
           >
             <div className="about-me">
-              <p className="leader">
-                Hi! I'm Shahid. I'm a software engineer with a keen eye for
-                design and a passion for fusing great design with beautiful
-                code.
-              </p>
-              <p>
-                I studied computer engineering at University of Ghana and I'm
-                currently working as a software engineer. I view my work as
-                continual education, and I'm constantly looking for chances to
-                collaborate with people who are as eager to offer their
-                knowledge as I am to absorb it.
-              </p>
-              <p>
-                My experience in web development extends from simple front-end
-                applications that utilize REST apis to full stack web projects
-                with a variety of tools and frameworks like PostgreSQL, HTML,
-                CSS, React and ASP.NET. User experience is a component of
-                software that, in my opinion, is essential for a successful
-                product.I'm incredibly passionate about UI effects, animations,
-                and developing dynamic, intuitive user interfaces.
-              </p>
-              <p>
-                I love discovering new and better ways to create seamless user
-                experiences with clean, efficient, and scalable code.My main
-                objective is to create something beautiful with people who bring
-                out the best in me, at the end of the day.
-              </p>
+              <div>
+                <p className="leader">
+                  Hi! I'm Shahid. I'm a software engineer with a keen eye for
+                  design and a passion for fusing great design with beautiful
+                  code.
+                </p>
+                <p>
+                  I studied computer engineering at University of Ghana and I'm
+                  currently working as a software engineer. I view my work as
+                  continual education, and I'm constantly looking for chances to
+                  collaborate with people who are as eager to offer their
+                  knowledge as I am to absorb it.
+                </p>
+
+                <p>
+                  My experience in web development extends from simple front-end
+                  applications that utilize REST apis to full stack web projects
+                  with a variety of tools and frameworks like PostgreSQL, HTML,
+                  CSS, React and ASP.NET. I'm incredibly passionate about UI
+                  effects, animations, and developing dynamic, intuitive user
+                  interfaces.
+                </p>
+                <p>
+                  I love discovering new and better ways to create seamless user
+                  experiences with clean, efficient, and scalable code.My main
+                  objective is to create something beautiful with people who
+                  bring out the best in me, at the end of the day.
+                </p>
+              </div>
             </div>
             <div className="">
               <div className="tagCloud"></div>
