+++ src/components/Sections/Contact/Contact.scss
@@ -1,0 +1,28 @@
+.contactSection {
+  .contact {
+    display: flex;
+    flex-wrap: wrap;
+    align-items: center;
+    justify-content: center;
+    .hello {
+      text-decoration: none;
+      border: solid 2px var(--accentColor);
+      color: var(--accentColor);
+      border-radius: 10px;
+      padding: 1.25rem 1.75rem;
+    }
+  }
+  footer {
+    padding-top: 90px;
+    .socials {
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      list-style: none;
+      column-gap: 0.25em;
+      li {
+        padding: 0 15px;
+      }
+    }
+  }
+}
