+++ src/components/Sections/Home/Home.scss
@@ -1,0 +1,3 @@
+.content {
+  text-align: center;
+}
