+++ src/components/Sections/Skills/Skills.scss
@@ -1,0 +1,39 @@
+.skills {
+  .chart {
+    footer {
+      background-color: #373737;
+      height: 2px;
+      div {
+        height: 100%;
+        animation: 1s ease-out 0s 1 slideInLeft;
+      }
+    }
+  }
+  .skillCard {
+    background-color: var(--darker-background);
+    border-radius: 10px;
+    color: var(--textShade200);
+    h5 {
+      font-size: var(--headingMd);
+      font-family: gulzar;
+      color: var(--textShade300);
+    }
+  }
+  .skillCard:hover {
+    transform: scale(1.05);
+    transition: transform 0.25s ease-out;
+    box-shadow: var(--hero-text-shadow);
+  }
+  .tag {
+    color: var(--textShade100);
+    font-size: var(--font-md);
+  }
+}
+@keyframes slideInLeft {
+  0% {
+    transform: translateX(0);
+  }
+  100% {
+    transform: translateX(100%);
+  }
+}
