--- src/components/Sections/Styles.scss
@@ -1,126 +1,0 @@
-.section {
-  .accentText {
-    color: var(--accentColor);
-    font-family: gulzar;
-  }
-  .content {
-    text-align: center;
-    .hero {
-      color: var(--textShade300);
-      font-size: var(--headingXXXL);
-      font-weight: 800;
-      font-family: gulzar;
-      text-shadow: var(--hero-text-shadow);
-    }
-  }
-  .profession {
-    font-size: clamp(2.7rem, 5vw, 4rem);
-    font-family: gulzar;
-    color: var(--textShade300);
-    text-shadow: 3px 3px rgba(0, 0, 0, 0.95);
-    font-weight: 600;
-  }
-  .heading {
-    font-size: var(--headingXL);
-    font-family: gulzar;
-    color: var(--textShade300);
-    font-weight: 600;
-  }
-  .subheading {
-    font-family: gulzar;
-    font-size: var(--headingL);
-    color: var(--textShade200);
-  }
-  .tag {
-    color: var(--textShade100);
-    font-size: var(--font-md);
-  }
-  .tagCloud {
-    color: var(--accentColor);
-    .tagcloud {
-      width: 100% !important;
-      bottom: 10%;
-    }
-  }
-
-  .about-me {
-    .leader {
-      font-size: var(--headingMd);
-      font-weight: 300;
-    }
-    p {
-      max-height: 200px;
-    }
-  }
-
-  .skills {
-    .chart {
-      footer {
-        background-color: #373737;
-        height: 2px;
-        div {
-          height: 100%;
-          animation: fadeInAnimation ease-out 3s;
-          animation-iteration-count: 1;
-          animation-fill-mode: forwards;
-        }
-      }
-    }
-    .skillCard {
-      background-color: var(--darker-background);
-      border-radius: 10px;
-      color: var(--textShade200);
-      h5 {
-        font-size: var(--headingMd);
-        font-family: gulzar;
-        color: var(--textShade300);
-      }
-    }
-    .skillCard:hover {
-      transform: scale(1.05);
-      transition: transform 0.25s ease-out;
-      box-shadow: var(--hero-text-shadow);
-    }
-  }
-}
-.contactSection {
-  .contact {
-    display: flex;
-    flex-wrap: wrap;
-    align-items: center;
-    justify-content: center;
-    .hello {
-      text-decoration: none;
-      border: solid 2px var(--accentColor);
-      color: var(--accentColor);
-      border-radius: 10px;
-      padding: 1.25rem 1.75rem;
-    }
-  }
-  footer {
-    padding-top: 90px;
-    .socials {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      list-style: none;
-      column-gap: 0.25em;
-      li {
-        padding: 0 15px;
-      }
-    }
-  }
-}
-@keyframes moveGradient {
-  50% {
-    background-position: 100% 50%;
-  }
-}
-@keyframes fadeInAnimation {
-  0% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
