+++ src/pages/Main.tsx
@@ -1,8 +1,8 @@
-import About from "../components/Sections/About";
-import Contact from "../components/Sections/Contact";
-import Home from "../components/Sections/Home";
-import Projects from "../components/Sections/Projects";
-import Skills from "../components/Sections/Skills";
+import About from "../components/Sections/About/About";
+import Contact from "../components/Sections/Contact/Contact";
+import Home from "../components/Sections/Home/Home";
+import Projects from "../components/Sections/Projects/Projects";
+import Skills from "../components/Sections/Skills/Skills";
 
 const Main: React.FC = () => {
   return (
